refactor(bind): add bind()/unbind() functions
Prior to this commit, binding a `blinker.Namespace` to a Paho MQTT
client was done by subclassing the `blinker.Namespace`. However, with
this approach, it was not possible to bind an _existing_ `Namespace`
instance to a MQTT client.

In this commit, introduce the `bind()` and `unbind()` functions, which
make it possible to dynamically manage binding between a blinker
`Namespace` and a Paho MQTT client.