fix(mqtt-proxy): allow subclassing
Prior to this commit, the `MqttProxy.from_uri()` class method did not
work as expected in classes inheriting from `MqttProxy`. Specifically, a
`MqttProxy` instance was returned, rather than an instance of the
subclass type.

In this commit, fix the `MqttProxy.from_uri()` class method to return an
instance of the subclass type instead.