feat(systray): add channels grid
Prior to this commit, the list of actuated channels was displayed in the
status window. However, there were two issues with this:

 1. The list text overflowed when there were many actuated channels.
 2. There was no way to toggle the actuation state of channels.

In this commit, add a grid displaying a button for each channel; where
the background color of each button corresponds to the actuation state
(green indicates the respective channel is actuated). In addition,
clicking on a button toggles the corresponding actuation state.